requirements_etl.txt
polars>=0.19
mysql-connector-python>=8.0
requests>=2.31
urllib3>=2.0
python-dotenv>=1.0
